Rlp module cleanup
