+++ basic_bootloader/src/bootloader/block_flow/ethereum/block_data.rs
@@ -1,5 +1,5 @@
+use crate::bootloader::block_flow::ethereum::mpt_leaf::CellEnvelope;
 use crate::bootloader::block_flow::ethereum::rlp_ordering_and_key_for_index;
-use crate::bootloader::rlp::CellEnvelope;
 use crate::bootloader::rlp::ReceiptEncoder;
 use crate::bootloader::transaction_flow::ethereum::EthereumTransactionFlow;
 use crate::bootloader::transaction_flow::ethereum::LogsBloom;
