+++ zk_ee/src/utils/write_bytes.rs
@@ -8,6 +8,12 @@ use crypto::MiniDigest;
 ///
 pub trait WriteBytes {
     fn write(&mut self, buf: &[u8]);
+
+    /// Writes a single byte. Defaults to a one-byte [`WriteBytes::write`], but
+    /// implementors can override it if they have a cheaper path.
+    fn write_byte(&mut self, byte: u8) {
+        self.write(&[byte]);
+    }
 }
 
 // implement it for hashers by default
